requirements.txt
watchdog
requests
python-dotenv
py7zr
aiohttp
aiofiles